separate status sheet from everything
